create a types folder
