+++ packages/two-factor/package.json
@@ -29,7 +29,9 @@
   },
   "dependencies": {
     "@accounts/types": "^0.1.0-beta.10",
+    "@types/lodash": "^4.14.108",
     "@types/speakeasy": "2.0.2",
+    "lodash": "^4.17.10",
     "speakeasy": "^2.0.0"
   },
   "devDependencies": {
