+++ packages/two-factor/src/index.ts
@@ -1,1 +1,3 @@
-export { TwoFactor, AccountsTwoFactorOptions } from './two-factor';
+export { TwoFactor } from './two-factor';
+export { AccountsTwoFactorOptions } from './types';
+export { getUserTwoFactorService} from './utils';
