+++ packages/two-factor/src/types/index.ts
@@ -1,0 +1,2 @@
+export { AccountsTwoFactorOptions } from './options';
+export { TwoFactorService } from './two-factor-service';
