+++ packages/two-factor/src/types/options.ts
@@ -1,0 +1,7 @@
+export interface AccountsTwoFactorOptions {
+  // Two factor app name
+  appName?: string;
+  // Two factor secret length, default to 20
+  secretLength?: number;
+  window?: number;
+}
