+++ packages/two-factor/src/types/two-factor-service.ts
@@ -1,0 +1,5 @@
+import * as speakeasy from 'speakeasy';
+
+export interface TwoFactorService {
+  secret: speakeasy.Key;
+}
