+++ packages/two-factor/src/utils/user.ts
@@ -1,0 +1,10 @@
+import { get } from 'lodash';
+import { User } from '@accounts/types';
+import { TwoFactorService } from '../types';
+
+/**
+ * Return the user two factor service object
+ */
+export const getUserTwoFactorService = (user: User): TwoFactorService | null => {
+  return get(user, ['services', 'two-factor'], null);
+};
